autocomply/requirements.txt
streamlit
sentence-transformers
scikit-learn
numpy
pandas
plotly
PyMuPDF
python-docx
anthropic
requests